+++ src/seppl/io/__init__.py
@@ -2,5 +2,5 @@ from ._reader import Reader
 from ._writer import Writer, StreamWriter, BatchWriter
 from ._filter import Filter, MultiFilter, FILTER_ACTIONS, FILTER_ACTION_KEEP, FILTER_ACTION_DISCARD
 from ._execution import execute
-from ._split import gcd, Splitter
+from ._split import gcd, Splitter, add_splitting_params, init_splitting_params, transfer_splitting_params, initialize_splitting
 from ._utils import locate_files